Split block-types core module responsibilities
