+++ .sampo/changesets/block-types-core-splits.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/block-types: patch
+---
+
+Split block support, variation, and binding internals by responsibility.
