+++ packages/wp-typia-block-types/src/blocks/bindings-core.ts
@@ -1,19 +1,31 @@
 import type { BlockAttributes } from "./registration.js";
 import {
-  DEFAULT_WORDPRESS_COMPATIBILITY_MIN_VERSION,
-  evaluateWordPressBlockApiCompatibility,
   type WordPressBlockApiCompatibilityDiagnostic,
   type WordPressBlockApiCompatibilityFeature,
   type WordPressBlockApiCompatibilityManifest,
   type WordPressCompatibilitySettings,
   type WordPressVersion,
 } from "./compatibility.js";
-import {
-  getDiagnosticSeverity,
-  handleDiagnostics,
-  type DiagnosticLogger,
-} from "./shared/diagnostics.js";
+import { type DiagnosticLogger } from "./shared/diagnostics.js";
 import { isObjectRecord } from "./shared/object-utils.js";
+import {
+  createBindingSourceDiagnostics,
+  handleBindingSourceDiagnostics,
+} from "./bindings-diagnostics.js";
+import {
+  collectBindingSourceCompatibilityFeatures,
+  createBindingCompatibilityManifest,
+  createBindingSourceCompatibilityManifest,
+} from "./bindings-manifest.js";
+import {
+  resolveDefineBindingSourceSettings,
+  splitDefineBindingSourceInput,
+} from "./bindings-settings.js";
+
+export {
+  collectBindingSourceCompatibilityFeatures,
+  createBindingSourceCompatibilityManifest,
+};
 
 export type BindingSourceName = `${string}/${string}`;
 
@@ -204,408 +216,6 @@ export type BindingSourceDiagnostic =
 export interface BindingSourceRegistrationEntry {
   readonly metadata: DefinedBlockBindingSourceMetadata;
   readonly source: DefinedBindingSource;
-}
-
-const DEFINE_BINDING_SOURCE_INLINE_OPTION_KEYS = new Set<string>([
-  "allowUnknownFutureKeys",
-  "editor",
-  "fieldsList",
-  "logger",
-  "minVersion",
-  "minWordPress",
-  "minWordPressEditor",
-  "minWordPressFieldsList",
-  "minWordPressServer",
-  "minWordPressSupportedAttributesFilter",
-  "onDiagnostic",
-  "server",
-  "strict",
-  "supportedAttributesFilter",
-]);
-
-const SOURCE_NAME_PATTERN =
-  /^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\/[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$/u;
-const FIELD_NAME_PATTERN = /^[A-Za-z_][A-Za-z0-9_-]*$/u;
-
-function isBindingSourceVersionGates(
-  value: WordPressVersion | BindingSourceVersionGates | undefined,
-): value is BindingSourceVersionGates {
-  return isObjectRecord(value);
-}
-
-function splitDefineBindingSourceInput<
-  TSource extends BindingSourceDefinition & DefineBindingSourceInlineOptions,
->(source: TSource): {
-  inlineOptions: DefineBindingSourceInlineOptions;
-  source: StripDefineBindingSourceOptions<TSource> & BindingSourceDefinition;
-} {
-  const inlineOptions: DefineBindingSourceInlineOptions = {};
-  const normalizedSource: Record<string, unknown> = {};
-
-  for (const [key, value] of Object.entries(source)) {
-    if (DEFINE_BINDING_SOURCE_INLINE_OPTION_KEYS.has(key)) {
-      Object.assign(inlineOptions, { [key]: value });
-      continue;
-    }
-
-    normalizedSource[key] = value;
-  }
-
-  return {
-    inlineOptions,
-    source: normalizedSource as StripDefineBindingSourceOptions<TSource> &
-      BindingSourceDefinition,
-  };
-}
-
-function resolveMinWordPress(
-  inlineOptions: DefineBindingSourceInlineOptions,
-  options: DefineBindingSourceOptions,
-): {
-  compatibility: WordPressCompatibilitySettings;
-  gates: BindingSourceVersionGates;
-} {
-  const optionMinWordPress = options.minWordPress;
-  const inlineMinWordPress = inlineOptions.minWordPress;
-  const optionGates = isBindingSourceVersionGates(optionMinWordPress)
-    ? optionMinWordPress
-    : {};
-  const inlineGates = isBindingSourceVersionGates(inlineMinWordPress)
-    ? inlineMinWordPress
-    : {};
-  const minVersion =
-    options.minVersion ??
-    (typeof optionMinWordPress === "string" ? optionMinWordPress : undefined) ??
-    inlineOptions.minVersion ??
-    (typeof inlineMinWordPress === "string" ? inlineMinWordPress : undefined);
-  const compatibility: WordPressCompatibilitySettings = {
-    strict: options.strict ?? inlineOptions.strict ?? true,
-  };
-
-  if (options.allowUnknownFutureKeys ?? inlineOptions.allowUnknownFutureKeys) {
-    Object.assign(compatibility, {
-      allowUnknownFutureKeys:
-        options.allowUnknownFutureKeys ?? inlineOptions.allowUnknownFutureKeys,
-    });
-  }
-  if (minVersion !== undefined) {
-    Object.assign(compatibility, { minVersion });
-  }
-
-  const gates: BindingSourceVersionGates = {};
-  const editor =
-    options.minWordPressEditor ??
-    optionGates.editor ??
-    inlineOptions.minWordPressEditor ??
-    inlineGates.editor;
-  const fieldsList =
-    options.minWordPressFieldsList ??
-    optionGates.fieldsList ??
-    inlineOptions.minWordPressFieldsList ??
-    inlineGates.fieldsList;
-  const server =
-    options.minWordPressServer ??
-    optionGates.server ??
-    inlineOptions.minWordPressServer ??
-    inlineGates.server;
-  const supportedAttributesFilter =
-    options.minWordPressSupportedAttributesFilter ??
-    optionGates.supportedAttributesFilter ??
-    inlineOptions.minWordPressSupportedAttributesFilter ??
-    inlineGates.supportedAttributesFilter;
-
-  if (editor !== undefined) {
-    Object.assign(gates, { editor });
-  }
-  if (fieldsList !== undefined) {
-    Object.assign(gates, { fieldsList });
-  }
-  if (server !== undefined) {
-    Object.assign(gates, { server });
-  }
-  if (supportedAttributesFilter !== undefined) {
-    Object.assign(gates, { supportedAttributesFilter });
-  }
-
-  return {
-    compatibility,
-    gates,
-  };
-}
-
-function resolveDefineBindingSourceSettings(
-  inlineOptions: DefineBindingSourceInlineOptions,
-  options: DefineBindingSourceOptions,
-  source: BindingSourceDefinition,
-): {
-  compatibility: WordPressCompatibilitySettings;
-  features: {
-    editor: boolean;
-    fieldsList: boolean;
-    metadata: boolean;
-    server: boolean;
-    supportedAttributesFilter: boolean;
-  };
-  gates: BindingSourceVersionGates;
-  logger: DefineBindingSourceOptions["logger"];
-  onDiagnostic: DefineBindingSourceOptions["onDiagnostic"];
-  strict: boolean;
-} {
-  const { compatibility, gates } = resolveMinWordPress(inlineOptions, options);
-  const strict = compatibility.strict ?? true;
-  const hasFields = (source.fields?.length ?? 0) > 0;
-  const hasBindableAttributes = (source.bindableAttributes?.length ?? 0) > 0;
-
-  return {
-    compatibility,
-    features: {
-      editor: options.editor ?? inlineOptions.editor ?? true,
-      fieldsList: options.fieldsList ?? inlineOptions.fieldsList ?? hasFields,
-      metadata: true,
-      server: options.server ?? inlineOptions.server ?? true,
-      supportedAttributesFilter:
-        options.supportedAttributesFilter ??
-        inlineOptions.supportedAttributesFilter ??
-        hasBindableAttributes,
-    },
-    gates,
-    logger: options.logger ?? inlineOptions.logger,
-    onDiagnostic: options.onDiagnostic ?? inlineOptions.onDiagnostic,
-    strict,
-  };
-}
-
-function getFeatureMinVersion(
-  feature: WordPressBlockApiCompatibilityFeature,
-  fallback: WordPressVersion,
-  gates: BindingSourceVersionGates,
-): WordPressVersion {
-  if (feature.area !== "blockBindings") {
-    return fallback;
-  }
-
-  switch (feature.feature) {
-    case "metadata.bindings":
-    case "serverRegistration":
-      return gates.server ?? fallback;
-    case "editorFieldsList":
-      return gates.fieldsList ?? fallback;
-    case "editorRegistration":
-    case "editorSourceLookup":
-      return gates.editor ?? fallback;
-    case "supportedAttributesFilter":
-      return gates.supportedAttributesFilter ?? gates.fieldsList ?? fallback;
-    default:
-      return fallback;
-  }
-}
-
-function createBindingCompatibilityManifest(
-  features: readonly WordPressBlockApiCompatibilityFeature[],
-  settings: WordPressCompatibilitySettings,
-  gates: BindingSourceVersionGates,
-): WordPressBlockApiCompatibilityManifest {
-  const fallback =
-    settings.minVersion ?? DEFAULT_WORDPRESS_COMPATIBILITY_MIN_VERSION;
-  const strict = settings.strict ?? true;
-  const allowUnknownFutureKeys = settings.allowUnknownFutureKeys ?? false;
-  const evaluations = features.map((feature) =>
-    evaluateWordPressBlockApiCompatibility(feature, {
-      allowUnknownFutureKeys,
-      minVersion: getFeatureMinVersion(feature, fallback, gates),
-      strict,
-    }),
-  );
-  const diagnostics = evaluations.flatMap((evaluation) =>
-    evaluation.diagnostic ? [evaluation.diagnostic] : [],
-  );
-
-  return {
-    allowUnknownFutureKeys,
-    diagnostics,
-    evaluations,
-    minVersion: fallback,
-    strict,
-    supported: evaluations.filter(
-      (evaluation) => evaluation.status === "supported",
-    ),
-    unknown: evaluations.filter((evaluation) => evaluation.status === "unknown"),
-    unsupported: evaluations.filter(
-      (evaluation) => evaluation.status === "unsupported",
-    ),
-  };
-}
-
-export function collectBindingSourceCompatibilityFeatures(
-  settings: {
-    readonly editor?: boolean;
-    readonly fieldsList?: boolean;
-    readonly metadata?: boolean;
-    readonly server?: boolean;
-    readonly supportedAttributesFilter?: boolean;
-  } = {},
-): readonly WordPressBlockApiCompatibilityFeature[] {
-  const features: WordPressBlockApiCompatibilityFeature[] = [];
-
-  if (settings.metadata ?? true) {
-    features.push({
-      area: "blockBindings",
-      feature: "metadata.bindings",
-    });
-  }
-  if (settings.server ?? true) {
-    features.push({
-      area: "blockBindings",
-      feature: "serverRegistration",
-    });
-  }
-  if (settings.editor ?? true) {
-    features.push({
-      area: "blockBindings",
-      feature: "editorRegistration",
-    });
-  }
-  if (settings.fieldsList ?? false) {
-    features.push({
-      area: "blockBindings",
-      feature: "editorFieldsList",
-    });
-  }
-  if (settings.supportedAttributesFilter ?? false) {
-    features.push({
-      area: "blockBindings",
-      feature: "supportedAttributesFilter",
-    });
-  }
-
-  return features;
-}
-
-export function createBindingSourceCompatibilityManifest(
-  settings: DefineBindingSourceOptions = {},
-): WordPressBlockApiCompatibilityManifest {
-  const resolved = resolveDefineBindingSourceSettings({}, settings, {
-    name: "wp-typia/binding-source",
-  });
-
-  return createBindingCompatibilityManifest(
-    collectBindingSourceCompatibilityFeatures(resolved.features),
-    resolved.compatibility,
-    resolved.gates,
-  );
-}
-
-function createBindingSourceDiagnostics(
-  source: BindingSourceDefinition,
-  options: {
-    readonly editor: boolean;
-    readonly fieldsList: boolean;
-    readonly server: boolean;
-    readonly strict: boolean;
-  },
-): readonly BindingSourceAuthoringDiagnostic[] {
-  const diagnostics: BindingSourceAuthoringDiagnostic[] = [];
-  const severity = getDiagnosticSeverity(options.strict);
-
-  if (!SOURCE_NAME_PATTERN.test(source.name)) {
-    diagnostics.push({
-      code: "invalid-source-name",
-      message: `Block binding source "${source.name}" must be lowercase and namespaced, such as "acme/profile-data".`,
-      severity,
-      sourceName: source.name,
-    });
-  }
-
-  if (options.server && !source.getValueCallback) {
-    diagnostics.push({
-      code: "missing-php-callback",
-      message: `Block binding source "${source.name}" needs getValueCallback when server registration is enabled.`,
-      severity,
-      sourceName: source.name,
-    });
-  }
-
-  if (options.fieldsList && !options.editor) {
-    diagnostics.push({
-      code: "fields-list-requires-editor",
-      message: `Block binding source "${source.name}" enables getFieldsList() without editor registration.`,
-      severity,
-      sourceName: source.name,
-    });
-  }
-
-  const seenFields = new Set<string>();
-  for (const field of source.fields ?? []) {
-    if (!FIELD_NAME_PATTERN.test(field.name)) {
-      diagnostics.push({
-        code: "invalid-field-name",
-        fieldName: field.name,
-        message: `Block binding source "${source.name}" field "${field.name}" must be a stable identifier.`,
-        severity,
-        sourceName: source.name,
-      });
-    }
-    if (seenFields.has(field.name)) {
-      diagnostics.push({
-        code: "duplicate-field-name",
-        fieldName: field.name,
-        message: `Block binding source "${source.name}" declares duplicate field "${field.name}".`,
-        severity,
-        sourceName: source.name,
-      });
-    }
-    seenFields.add(field.name);
-  }
-
-  for (const target of source.bindableAttributes ?? []) {
-    if (!SOURCE_NAME_PATTERN.test(target.blockName)) {
-      diagnostics.push({
-        blockName: target.blockName,
-        code: "invalid-block-name",
-        message: `Bindable attributes target "${target.blockName}" must be a lowercase namespaced block name.`,
-        severity,
-        sourceName: source.name,
-      });
-    }
-
-    const seenAttributes = new Set<string>();
-    for (const attribute of target.attributes) {
-      if (!FIELD_NAME_PATTERN.test(attribute)) {
-        diagnostics.push({
-          attribute,
-          blockName: target.blockName,
-          code: "invalid-bindable-attribute",
-          message: `Bindable attribute "${attribute}" for "${target.blockName}" must be a stable identifier.`,
-          severity,
-          sourceName: source.name,
-        });
-      }
-      if (seenAttributes.has(attribute)) {
-        diagnostics.push({
-          attribute,
-          blockName: target.blockName,
-          code: "duplicate-bindable-attribute",
-          message: `Bindable attribute "${attribute}" for "${target.blockName}" is declared more than once.`,
-          severity,
-          sourceName: source.name,
-        });
-      }
-      seenAttributes.add(attribute);
-    }
-  }
-
-  return diagnostics;
-}
-
-function handleBindingSourceDiagnostics(
-  diagnostics: readonly BindingSourceDiagnostic[],
-  onDiagnostic: DefineBindingSourceOptions["onDiagnostic"],
-  logger: DefineBindingSourceOptions["logger"],
-): void {
-  handleDiagnostics(diagnostics, onDiagnostic, {
-    failureHeading: "WordPress block binding source check failed:",
-    logger,
-  });
 }
 
 export function getDefinedBindingSourceMetadata(
